+++ backend/tests/baserow/contrib/database/field/test_field_types.py
@@ -98,10 +98,15 @@ def test_url_field_type(data_fixture):
             user=user, table=table, values={"url": "httpss"}, model=model
         )
 
+    very_long_url = "https://baserow.io/with-a-very-long-url-that-exceeds-the-old-254-"
+    "char-limit/with-a-very-long-url-that-exceeds-the-old-254-char-limit/with-a-very-"
+    "long-url-that-exceeds-the-old-254-char-limit/with-a-very-long-url-that-exceeds"
+    "-the-old-254-char/with-a-very-long-url-that-exceeds-the-old-254-char/"
+
     row_handler.create_row(
         user=user,
         table=table,
-        values={"name": "http://test.nl", "url": "https://baserow.io", "number": 5},
+        values={"name": "http://test.nl", "url": very_long_url, "number": 5},
         model=model,
     )
     row_handler.create_row(
@@ -153,7 +158,7 @@ def test_url_field_type(data_fixture):
     rows = model.objects.all()
 
     assert rows[0].name == "http://test.nl"
-    assert rows[0].url == "https://baserow.io"
+    assert rows[0].url == very_long_url
     assert rows[0].number == ""
 
     assert rows[1].name == ""